Refactor: Emoji 관련 함수 HOOK으로 빼기
