+++ src/modules/quiz/components/multi/UserTags/UserTags.tsx
@@ -1,48 +1,21 @@
-import { useState, useRef } from "react";
+import { useRef } from "react";
 import { TeamUserTagProps } from "../../../../../types/room";
 import { UserTag, UserTagImg, UserTagsContainer } from "./styled";
 import { useTeamState } from "../../../../../contexts/TeamStateContext/useTeamState";
 import { useGameUser } from "../../../../../contexts/GameUserContext/useGameUser";
 import { EmojiButton } from "../../../../../components/buttons/emoji";
 import { AnimatedEmoji } from "../../../../../components/modal/emogi/animatedEmoji";
+import { useEmoji } from "../../../../../hook/useEmoji";
 
 export const UserTagsComponent = ({ teamId }: TeamUserTagProps) => {
     const { user } = useGameUser();
     const { teamOneUsers, teamTwoUsers } = useTeamState()
     const teamUsers = teamId == 1 ? teamOneUsers : teamTwoUsers;
 
-    // 각 유저의 이모티콘 상태 관리
-    const [animatedEmojis, setAnimatedEmojis] = useState<Array<{
-        id: number;
-        src: string;
-        x: number;
-        y: number;
-    }>>([]);
-
     // UserTag의 ref를 저장할 객체
     const userTagRefs = useRef<{ [key: string]: HTMLDivElement | null }>({});
 
-    const handleEmojiSelect = (emojiPath: string, targetUsername: string) => {
-        const targetTagElement = userTagRefs.current[targetUsername];
-        if (targetTagElement) {
-            // 부모 컨테이너(UserTagsContainer)의 위치 가져오기
-            const containerRect = targetTagElement.parentElement?.getBoundingClientRect();
-            const tagRect = targetTagElement.getBoundingClientRect();
-
-            if (containerRect) {
-                // 상대적 위치 계산
-                const relativeX = tagRect.left - containerRect.left;
-                const relativeY = tagRect.top - containerRect.top;
-
-                setAnimatedEmojis(prev => [...prev, {
-                    id: Date.now(),
-                    src: emojiPath,
-                    x: relativeX + (tagRect.width / 2), // 태그의 중앙
-                    y: relativeY - 100 // 태그 위
-                }]);
-            }
-        }
-    };
+    const { animatedEmojis, handleEmojiSelect } = useEmoji(userTagRefs);
 
     return (
         <UserTagsContainer>
